// frontend/src/app/api/notes/model/createNoteDto.ts
/**
 * PaperNote Notes API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request model for creating a new note with title, content, optional tags and initial sharing configuration
 */
export interface CreateNoteDto { 
    title?: string | null;
    content?: string | null;
    tags?: Array<string> | null;
    sharedWithUsernames?: Array<string> | null;
}