// frontend/src/app/api/notes/model/updateNoteDto.ts
/**
 * PaperNote Notes API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request model for updating an existing note\'s content, tags and sharing configuration
 */
export interface UpdateNoteDto { 
    id?: string;
    title?: string | null;
    content?: string | null;
    tags?: Array<string> | null;
    sharedWithUsernames?: Array<string> | null;
}